# server/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Pomodro Backend API",
      "version": "1.0.0",
      "description": "API documentation for the Pomodro Backend Server."
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Local server"
      }
    ],
    "paths": {
...
      "/logs": {
        "get": {
          "summary": "Visualizes logs with filtering and pagination.",
          "description": "Optional query parameters: log_type, filename, page, per_page.",
          "parameters": [
            {
              "name": "log_type",
              "in": "query",
              "schema": { "type": "string" },
              "description": "Filter logs by type"
            },
            {
              "name": "filename",
              "in": "query",
              "schema": { "type": "string" },
              "description": "Filter logs by filename"
            },
            {
              "name": "page",
              "in": "query",
              "schema": { "type": "integer" },
              "description": "Page number for pagination"
            },
            {
              "name": "per_page",
              "in": "query",
              "schema": { "type": "integer" },
              "description": "Number of logs per page"
            }
          ],
          "responses": {
            "200": { "description": "Logs returned successfully." }
          }
        }
      },
      "/logs/trend": {
        "get": {
          "summary": "Displays the trend chart of logs.",
          "responses": {
            "200": { "description": "Trend chart displayed successfully." },
            "404": { "description": "No valid logs available." }
          }
        }
      }
    },
    "components": {
      "securitySchemes": {
        "JWT": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "JWT": []
      }
    ]
  }